pet_stmt_build_ast_exprs: handle some nested operations inside accesses
In particular, support additions and subtractions,
as well as integer constants.
These are handled by calling the corresponding operation
on objects of type isl_ast_expr.

The same mechanism could be used for other operations,
but only if there is a corresponding isl_ast_expr operation.

For operations that do not have a corresponding isl_ast_expr operation,
a different mechanism would have to be used.
One possibility would be to leave in the identifier
holding a pointer to an isl_ast_expr and then to introduce
an isl_ast_expr_print in isl taking an isl_ast_print_options object
with a callback for printing identifiers.

Requested-by: Emil Vatai <emil.vatai@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>