allow extraction of function summary to fail
If so, do not try to extract more accurate access relations
from the called function and simply use the default heuristics instead.

This is illustrated by the two new test cases.
In call9, the called function can be analyzed and
the access relations in the caller can be updated
to reflect that only odd elements are being accessed.
In call10, the called function cannot be analyzed
(because it contains a goto) and this extra information
is not available in the access relations of the caller.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>